Add phase for applying a cluster
